+++ packages/terra-functional-testing/src/constants/index.js
@@ -25,4 +25,10 @@ const TERRA_VIEWPORTS = {
   },
 };
 
-module.exports = { TERRA_VIEWPORTS };
+const BUILD_BRANCH = {
+  master: 'master',
+  dev: 'dev',
+  pullRequest: /pr-\d+/g,
+};
+
+module.exports = { TERRA_VIEWPORTS, BUILD_BRANCH };