+++ packages/terra-functional-testing/src/terra-cli/wdio/test-runner.js
@@ -3,6 +3,7 @@ const path = require('path');
 const Launcher = require('@wdio/cli').default;
 const { Logger } = require('@cerner/terra-cli');
 const getConfigurationOptions = require('../../config/utils/getConfigurationOptions');
+const getRemoteScreenshotConfiguration = require('../../config/utils/getRemoteScreenshotConfiguration');
 const { cleanScreenshots, downloadScreenshots } = require('../../commands/utils');
 
 const logger = new Logger({ prefix: '[terra-functional-testing:wdio]' });
@@ -33,6 +34,7 @@ class TestRunner {
    * @param {Object} options - The test run options.
    * @param {number} options.assetServerPort - The port to run the webpack and express asset services on.
    * @param {string} options.browsers - A list of browsers for the test run.
+   * @param {string} options.buildBranch - The type of branch being built.
    * @param {string} options.config - A file path to the test runner configuration.
    * @param {boolean} options.disableSeleniumService - A flag to disable the selenium docker service.
    * @param {string} options.externalHost - The host address the testing environment is connected to.
@@ -77,6 +79,7 @@ class TestRunner {
    * @param {Object} options - The test run options.
    * @param {number} options.assetServerPort - The port to run the webpack and express asset services on.
    * @param {string} options.browsers - A list of browsers for the test run.
+   * @param {string} options.buildBranch - The type of branch being built.
    * @param {string} options.config - A file path to the test runner configuration.
    * @param {boolean} options.disableSeleniumService - A flag to disable the selenium docker service.
    * @param {string} options.externalHost - The host address the testing environment is connected to.
@@ -103,10 +106,12 @@ class TestRunner {
       ...cliOptions
     } = options;
 
-    // Clean only the non reference screenshots.
-    cleanScreenshots();
+    cleanScreenshots(options.useRemoteReferenceScreenshots);
 
-    await TestRunner.configureScreenshots(options);
+    if (options.useRemoteReferenceScreenshots) {
+      const screenshotConfig = getRemoteScreenshotConfiguration();
+      await downloadScreenshots(screenshotConfig);
+    }
 
     /**
      * The following code loops through each permutation of theme, locale, and form factor.
@@ -133,31 +138,6 @@ class TestRunner {
           formFactorIndex += 1;
         } while (formFactorIndex < formFactors.length);
       }
-    }
-  }
-
-  static async configureScreenshots(options) {
-    const {
-      config,
-      screenshotUrl,
-    } = options;
-
-    let url = screenshotUrl;
-
-    if (!url) {
-      const configPath = TestRunner.configPath(config);
-      // eslint-disable-next-line global-require, import/no-dynamic-require
-      const wdioConfig = require(configPath);
-      const { screenshots } = wdioConfig.config;
-
-      if (screenshots) {
-        url = screenshots.url;
-      }
-    }
-
-    if (url) {
-      logger.info(`Starting to download screenshots from ${url}.`);
-      await downloadScreenshots(url);
     }
   }
 }